Common BaseObject Class.
