+++ Source/AliveLibAO/BaseGameObject.hpp
@@ -2,6 +2,7 @@
 
 #include "../AliveLibCommon/FunctionFwd.hpp"
 #include "../AliveLibCommon/BitField.hpp"
+#include "../AliveLibCommon/BaseGameObjectCommon.hpp"
 #include "DynamicArray.hpp"
 
 void AliveLibAO_ForceLink();
@@ -117,20 +118,6 @@ enum class Types : s16
 class BaseGameObject
 {
 public:
-    enum Options
-    {
-        eListAddFailed_Bit1 = 0x001,
-        eUpdatable_Bit2 = 0x002,
-        eDead_Bit3 = 0x004,
-        eDrawable_Bit4 = 0x008,
-        eIsBaseAnimatedWithPhysicsObj_Bit5 = 0x010,
-        eIsBaseAliveGameObject_Bit6 = 0x020,
-        eCanExplode_Bit7 = 0x040,
-        eInteractive_Bit8 = 0x080,
-        eSurviveDeathReset_Bit9 = 0x100,
-        eUpdateDuringCamSwap_Bit10 = 0x200,
-        eCantKill_Bit11 = 0x400
-    };
 
     EXPORT BaseGameObject* ctor_487E10(s16 arraySize);
     EXPORT BaseGameObject* dtor_487DF0();
@@ -162,7 +149,7 @@ private:
 
 public:
     Types field_4_typeId;
-    BitField16<Options> field_6_flags;
+    BitField16<BaseGameObjectCommon::Options> field_6_flags;
     s32 field_8_update_delay;
     s8 field_C_refCount;
     s8 field_D_padding;
